avm2: Store interpreter instruction pointer as usize instead of i32
We also now store the absolute position to branch to in branch ops instead of storing the offset.

+++ core/src/avm2/optimizer/optimize.rs
@@ -533,7 +533,7 @@ pub fn optimize<'gc>(
     code: &mut Vec<Op<'gc>>,
     resolved_parameters: &[ResolvedParamConfig<'gc>],
     method_exceptions: &[Exception<'gc>],
-    jump_targets: &HashSet<i32>,
+    jump_targets: &HashSet<usize>,
 ) -> Result<(), Error<'gc>> {
     // These make the code less readable
     #![allow(clippy::collapsible_if)]
